+++ src/CommunityPatchLauncher/UserControls/LaunchGameUserControl.xaml.cs
@@ -2,6 +2,7 @@
 using CommunityPatchLauncher.Enums;
 using CommunityPatchLauncher.ViewModels;
 using System;
+using System.Windows;
 using System.Windows.Controls;
 
 namespace CommunityPatchLauncher.UserControls
@@ -11,10 +12,10 @@ namespace CommunityPatchLauncher.UserControls
     /// </summary>
     public partial class LaunchGameUserControl : UserControl, IDisposable
     {
-        public LaunchGameUserControl()
+        public LaunchGameUserControl(Window mainWindow)
         {
             InitializeComponent();
-            DataContext = new LaunchGameModelView(this);
+            DataContext = new LaunchGameModelView(this, mainWindow);
         }
 
         public void Dispose()